Add unittests for pushbullet.py
Current tests in pushbullet.py are integration tests. Unittesting is therefore required for local development without relying on the API service. This commit introduces the basic layout of the unittests with helper method for mocking the requests methods and returning json responses based off the API documentation on the pushbullet website.