Edit mock_response for nickname testing and added more tests
mock_response needed to account for when values are sent in JSON as None or empty, to behave more realistically like the pushbullet API. The pushbullet API removes these fields from the returned response.

Added fuller tests around nickname and added some future tests for yet to be added functionality.